+++ module/macros/macros.js
@@ -1,11 +1,6 @@
 import {
-  renameTag
-} from "./macro-helper.js";
-import {
   spendEffortString,
-  calculateAttackDifficultyString,
-  renameTagString,
-  itemMacroString
+  calculateAttackDifficultyString
 } from "./macro-strings.js";
 import {
   chatCardProposeIntrusion,
@@ -260,7 +255,7 @@ export async function allInOneRollDialog(actor, pool, skill, assets, effort1, ef
   });
 }
 
-export async function itemRollMacro(actor, itemID, pool, skillLevel, assets, effort1, effort2, additionalSteps, additionalCost, damage, effort3, damagePerLOE, teen, stepModifier, noRoll, bonus) {
+export async function itemRollMacro(actor, itemID, pool, skillLevel, assets, effort1, effort2, additionalSteps, additionalCost, damage, effort3, damagePerLOE, teen, stepModifier, noRoll, bonus, macroUuid) {
   // Find actor based on item ID
   const owner = game.actors.find(actor => actor.items.get(itemID));
 
@@ -341,6 +336,7 @@ export async function itemRollMacro(actor, itemID, pool, skillLevel, assets, eff
   if (!damagePerLOE) damagePerLOE = item.system.settings.rollButton.damagePerLOE;
   if (!teen) teen = (actor.system.basic.unmaskedForm == "Teen") ? true : false;
   if (!bonus) bonus = item.system.settings.rollButton.bonus;
+  if (!macroUuid) macroUuid = item.system.settings.rollButton.macroUuid;
 
   // Create item type
   let itemType = "";
@@ -358,6 +354,7 @@ export async function itemRollMacro(actor, itemID, pool, skillLevel, assets, eff
   rollEngineMain({
     actorUuid: actor.uuid,
     itemID: itemID,
+    macroUuid: macroUuid,
     teen: teen,
     skipRoll: noRoll,
     initiative: initiativeRoll,