+++ module/utilities/roll-engine/roll-engine-main.js
@@ -6,6 +6,7 @@ export async function rollEngineMain(data) {
   data = Object.assign({
     actorUuid: undefined,
     itemID: undefined,
+    macroUuid: undefined,
     teen: undefined,
     skipDialog: !game.settings.get("cyphersystem", "itemMacrosUseAllInOne"),
     skipRoll: false,
@@ -88,6 +89,12 @@ export async function rollEngineMain(data) {
   if (data.skillLevel == "Trained") data.skillLevel = 1;
   if (data.skillLevel == "Practiced") data.skillLevel = 0;
   if (data.skillLevel == "Inability") data.skillLevel = -1;
+
+  // Check for macro
+  if (data.macroUuid) {
+    let macro = fromUuidSync(data.macroUuid);
+    if (!macro) return ui.notifications.warn(game.i18n.localize("CYPHERSYSTEM.MacroNotFound"));
+  }
 
   // Go to the next step after checking whether dialog should be skipped
   if (!data.skipDialog) {