+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [unreleased]
+## [2.14.0] - 2014-05-18
 ### Added
 - Items with roll buttons now can execute macros when rolling. Specify a macro uuid (right-click the ID button in the header of the item sheet) in the item settings and it will be executed when a roll is made. `rollData` provides all the data of the roll for use in the macro.
 - Tags and recursions can also execeute macros when they are enabled. It works the same as for rolls, but the data is provided as `taggingData` to the macros.